updated product tests
