+++ tests/pw/pages/vendorPage.ts
@@ -48,7 +48,7 @@ export class VendorPage extends BasePage {
 		await this.searchProduct(productName);
 		await this.hover(selector.vendor.product.productCell(productName));
 		await this.clickAndWaitForResponseAndLoadState(data.subUrls.frontend.vDashboard.products, selector.vendor.product.editProduct(productName));
-
+		await this.toHaveValue(selector.vendor.product.edit.title, productName);
 	}
 
 
@@ -62,7 +62,7 @@ export class VendorPage extends BasePage {
 		await this.clearAndType(selector.vendor.vRegistration.regEmail, username + data.vendor.vendorInfo.emailDomain);
 		await this.clearAndType(selector.vendor.vRegistration.regPassword, vendorInfo.password);
 		await this.focusAndClick(selector.vendor.vRegistration.regVendor);
-		// await this.waitForVisibleLocator(selector.vendor.vRegistration.firstName);
+		await this.waitForVisibleLocator(selector.vendor.vRegistration.firstName);
 		await this.clearAndType(selector.vendor.vRegistration.firstName, username);
 		await this.clearAndType(selector.vendor.vRegistration.lastName, vendorInfo.lastName());
 		await this.clearAndType(selector.vendor.vRegistration.shopName, vendorInfo.shopName);
@@ -173,6 +173,17 @@ export class VendorPage extends BasePage {
 	}
 
 
+	// vendor account details render properly
+	async vendorAccountDetailsRenderProperly(){
+		await this.goIfNotThere(data.subUrls.frontend.vDashboard.editAccountVendor);
+
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
+		const { saveSuccessMessage, ...accountDetails } = selector.vendor.vAccountDetails;
+		await this.multipleElementVisible(accountDetails);
+
+	}
+
+
 	// vendor add vendor details
 	async addVendorDetails(vendor: vendor): Promise<void> {
 		await this.goIfNotThere(data.subUrls.frontend.vDashboard.editAccountVendor);
@@ -187,6 +198,7 @@ export class VendorPage extends BasePage {
 		// cleanup: reset password
 		// await this.updatePassword(vendor.vendorInfo.password1, vendor.vendorInfo.password, true);
 	}
+
 
 	// vendor update password
 	async updatePassword(currentPassword: string, newPassword: string, saveChanges = false): Promise<void> {
@@ -203,138 +215,8 @@ export class VendorPage extends BasePage {
 	//todo: fixed above functions
 
 
-	// products
-
-	// vendor add simple product
-	async vendorAddSimpleProduct(product: product['simple'] | product['simpleSubscription'] | product['external']): Promise<void> {
-		await this.goIfNotThere(data.subUrls.frontend.vDashboard.products);
-		const productName = product.productName();
-		// add new simple product
-		await this.click(selector.vendor.product.addNewProduct);
-		await this.waitForVisibleLocator(selector.vendor.product.productName);
-		await this.type(selector.vendor.product.productName, productName);
-		await this.type(selector.vendor.product.productPrice, product.regularPrice());
-		await this.addProductCategory(product.category); //todo:  split in separate test
-		await this.clickAndWaitForResponse(data.subUrls.ajax, selector.vendor.product.createProduct);
-		const createdProduct = await this.getElementValue(selector.vendor.product.title);
-		expect(createdProduct.toLowerCase()).toBe(productName.toLowerCase());
-	}
 
 
-	// vendor add variable product
-	async vendorAddVariableProduct(product: product['variable'] ): Promise<void> {
-		await this.vendorAddSimpleProduct(product);
-
-		// edit product
-		await this.selectByValue(selector.vendor.product.productType, product.productType);
-		// add variation
-		await this.selectByValue(selector.vendor.product.attribute.customProductAttribute, `pa_${product.attribute}`);
-		await this.click(selector.vendor.product.attribute.addAttribute);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.selectAll);
-		await this.click(selector.vendor.product.attribute.selectAll);
-		await this.click(selector.vendor.product.attribute.usedForVariations);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.saveAttributes);
-		await this.click(selector.vendor.product.attribute.saveAttributes);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.addVariations);
-		await this.selectByValue(selector.vendor.product.attribute.addVariations, product.variations.linkAllVariation);
-		await this.click(selector.vendor.product.attribute.go);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.confirmGo);
-		await this.click(selector.vendor.product.attribute.confirmGo);
-		await this.click(selector.vendor.product.attribute.okSuccessAlertGo);
-		await this.selectByValue(selector.vendor.product.attribute.addVariations, product.variations.variableRegularPrice);
-		await this.click(selector.vendor.product.attribute.go);
-		// await this.waitForVisibleLocator(selector.vendor.product.attribute.variationPrice)
-		await this.type(selector.vendor.product.attribute.variationPrice, product.regularPrice());
-		await this.click(selector.vendor.product.attribute.okVariationPrice);
-		await this.clickAndWaitForResponseAndLoadState(data.subUrls.frontend.vDashboard.products, selector.vendor.product.saveProduct, 302);
-		// const productCreateSuccessMessage = await this.getElementText(selector.vendor.product.updatedSuccessMessage);
-		// expect(productCreateSuccessMessage?.replace(/\s+/g, ' ').trim()).toMatch(product.saveSuccessMessage);
-		await this.toContainText(selector.vendor.product.updatedSuccessMessage, product.saveSuccessMessage);
-	}
-
-
-	// vendor add simple subscription product
-	async vendorAddSimpleSubscription(product: product['simpleSubscription']): Promise<void> {
-		await this.vendorAddSimpleProduct(product);
-		// edit product
-		await this.selectByValue(selector.vendor.product.productType, product.productType);
-		await this.type(selector.vendor.product.subscriptionPrice, product.subscriptionPrice());
-		await this.selectByValue(selector.vendor.product.subscriptionPeriodInterval, product.subscriptionPeriodInterval);
-		await this.selectByValue(selector.vendor.product.subscriptionPeriod, product.subscriptionPeriod);
-		await this.selectByValue(selector.vendor.product.expireAfter, product.expireAfter);
-		await this.type(selector.vendor.product.subscriptionTrialLength, product.subscriptionTrialLength);
-		await this.selectByValue(selector.vendor.product.subscriptionTrialPeriod, product.subscriptionTrialPeriod);
-		await this.clickAndWaitForResponseAndLoadState(data.subUrls.frontend.vDashboard.products, selector.vendor.product.saveProduct, 302);
-		// const productCreateSuccessMessage = await this.getElementText(selector.vendor.product.updatedSuccessMessage);
-		// expect(productCreateSuccessMessage?.replace(/\s+/g, ' ').trim()).toMatch(product.saveSuccessMessage);
-		await this.toContainText(selector.vendor.product.updatedSuccessMessage, product.saveSuccessMessage);
-	}
-
-
-	// vendor add variable subscription product
-	async vendorAddVariableSubscription(product: product['variableSubscription']): Promise<void> {
-		await this.vendorAddSimpleProduct(product);
-		// edit product
-		await this.selectByValue(selector.vendor.product.productType, product.productType);
-		// add variation
-		await this.selectByValue(selector.vendor.product.attribute.customProductAttribute, `pa_${product.attribute}`);
-		await this.click(selector.vendor.product.attribute.addAttribute);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.selectAll);
-		await this.click(selector.vendor.product.attribute.selectAll);
-		await this.click(selector.vendor.product.attribute.usedForVariations);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.saveAttributes);
-		await this.click(selector.vendor.product.attribute.saveAttributes);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.addVariations);
-		await this.selectByValue(selector.vendor.product.attribute.addVariations, product.variations.linkAllVariation);
-		await this.click(selector.vendor.product.attribute.go);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.confirmGo);
-		await this.click(selector.vendor.product.attribute.confirmGo);
-		await this.click(selector.vendor.product.attribute.okSuccessAlertGo);
-		await this.selectByValue(selector.vendor.product.attribute.addVariations, product.variations.variableRegularPrice);
-		await this.click(selector.vendor.product.attribute.go);
-		await this.waitForVisibleLocator(selector.vendor.product.attribute.variationPrice);
-		await this.type(selector.vendor.product.attribute.variationPrice, product.regularPrice());
-		await this.click(selector.vendor.product.attribute.okVariationPrice); // todo : add waitForResponse with click
-		await this.clickAndWaitForResponseAndLoadState(data.subUrls.frontend.vDashboard.products, selector.vendor.product.saveProduct, 302);
-
-		await this.waitForVisibleLocator(selector.vendor.product.updatedSuccessMessage);
-		// const productCreateSuccessMessage = await this.getElementText(selector.vendor.product.updatedSuccessMessage);
-		// expect(productCreateSuccessMessage?.replace(/\s+/g, ' ').trim()).toMatch(product.saveSuccessMessage);
-		await this.toContainText(selector.vendor.product.updatedSuccessMessage, product.saveSuccessMessage);
-	}
-
-
-	// vendor add external product
-	async vendorAddExternalProduct(product: product['external']): Promise<void> {
-		await this.vendorAddSimpleProduct(product);
-		// edit product
-		await this.selectByValue(selector.vendor.product.productType, product.productType);
-		await this.type(selector.vendor.product.productUrl, this.getBaseUrl() + product.productUrl);
-		await this.type(selector.vendor.product.buttonText, product.buttonText);
-		await this.clearAndType(selector.vendor.product.price, product.regularPrice());
-
-		await this.clickAndWaitForResponseAndLoadState(data.subUrls.frontend.vDashboard.products, selector.vendor.product.saveProduct, 302);
-		// const productCreateSuccessMessage = await this.getElementText(selector.vendor.product.updatedSuccessMessage);
-		// expect(productCreateSuccessMessage?.replace(/\s+/g, ' ').trim()).toMatch(product.saveSuccessMessage);
-		await this.toContainText(selector.vendor.product.updatedSuccessMessage, product.saveSuccessMessage);
-	}
-
-
-	// vendor add product category
-	async addProductCategory(category: string): Promise<void> {
-		await this.click(selector.vendor.product.productCategoryModal);
-		await this.waitForVisibleLocator(selector.vendor.product.productCategorySearchInput);
-		await this.type(selector.vendor.product.productCategorySearchInput, category);
-		await this.click(selector.vendor.product.productCategorySearchResult);
-		await this.click(selector.vendor.product.productCategoryDone);
-
-		const categoryAlreadySelectedPopup = await this.isVisible(selector.vendor.product.productCategoryAlreadySelectedPopup);
-		if (categoryAlreadySelectedPopup) {
-			await this.click(selector.vendor.product.productCategoryAlreadySelectedPopup);
-			await this.click(selector.vendor.product.productCategoryModalClose);
-		}
-		//todo:  handle multiple category selection with assertion
-	}
 
 
 	// product update
@@ -411,6 +293,7 @@ export class VendorPage extends BasePage {
 
 	// vendor settings
 
+	// update store map via settings save
 	async updateStoreMapViaSettingsSave() {
 		await this.goIfNotThere(data.subUrls.frontend.vDashboard.settingsStore);
 		await this.clickAndWaitForResponseAndLoadState(data.subUrls.ajax, selector.vendor.vStoreSettings.updateSettings);
@@ -727,130 +610,6 @@ export class VendorPage extends BasePage {
 	}
 
 
-	// vendor set verification settings
-	async setVerificationSettings(verification: vendor['verification']): Promise<void> {
-		await this.sendIdVerificationRequest(verification);
-		await this.sendAddressVerificationRequest(verification);
-		await this.sendCompanyVerificationRequest(verification);
-	}
-
-
-
-
-
-	// vendor shipping settings
-
-	// vendor set all shipping settings
-	// async setAllShippingSettings(): Promise<void> {
-	// 	await this.setShippingSettings(data.vendor.shipping.shippingMethods.flatRate);
-	// 	await this.setShippingSettings(data.vendor.shipping.shippingMethods.freeShipping);
-	// 	await this.setShippingSettings(data.vendor.shipping.shippingMethods.localPickup);
-	// 	await this.setShippingSettings(data.vendor.shipping.shippingMethods.tableRateShipping);
-	// 	await this.setShippingSettings(data.vendor.shipping.shippingMethods.distanceRateShipping);
-	// }
-
-
-	// set shipping policies
-	async setShippingPolicies(shippingPolicy: vendor['shipping']['shippingPolicy']): Promise<void> {
-		await this.goIfNotThere(data.subUrls.frontend.vDashboard.settingsShipping);
-		await this.click(selector.vendor.vShippingSettings.shippingPolicies.clickHereToAddShippingPolicies);
-		await this.selectByValue(selector.vendor.vShippingSettings.shippingPolicies.processingTime, shippingPolicy.processingTime);
-		await this.clearAndType(selector.vendor.vShippingSettings.shippingPolicies.shippingPolicy, shippingPolicy.shippingPolicy);
-		await this.type(selector.vendor.vShippingSettings.shippingPolicies.refundPolicy, shippingPolicy.refundPolicy);
-		await this.click(selector.vendor.vShippingSettings.shippingPolicies.saveSettings);
-		await this.toContainText(selector.vendor.vShippingSettings.updateSettingsSuccessMessage, shippingPolicy.saveSuccessMessage);
-	}
-
-
-	// vendor set shipping settings
-	async setShippingSettings(shipping: any): Promise<void> {
-		await this.goIfNotThere(data.subUrls.frontend.vDashboard.settingsShipping);
-		// edit shipping zone
-		await this.hover(selector.vendor.vShippingSettings.shippingZoneCell(shipping.shippingZone));
-		await this.clickAndWaitForResponse(data.subUrls.ajax, selector.vendor.vShippingSettings.editShippingZone(shipping.shippingZone));
-
-		// add shipping method if not available
-		const methodIsVisible = await this.isVisible(selector.vendor.vShippingSettings.shippingMethodCell(shipping.shippingMethod));
-		if (!methodIsVisible) {
-			await this.click(selector.vendor.vShippingSettings.addShippingMethod);
-			await this.selectByValue(selector.vendor.vShippingSettings.shippingMethod, shipping.selectShippingMethod);
-			await this.clickAndWaitForResponse(data.subUrls.ajax, selector.vendor.vShippingSettings.shippingMethodPopupAddShippingMethod);
-			await expect(this.page.getByText(shipping.shippingMethodSaveSuccessMessage)).toBeVisible();
-			await expect(this.page.getByText(shipping.zoneSaveSuccessMessage)).toBeVisible();
-		}
-		// edit shipping method
-		await this.hover(selector.vendor.vShippingSettings.shippingMethodCell(shipping.shippingMethod));
-		await this.click(selector.vendor.vShippingSettings.editShippingMethod(shipping.shippingMethod));
-
-		switch (shipping.selectShippingMethod) {
-		// flat rate
-		case 'flat_rate' :
-			await this.clearAndType(selector.vendor.vShippingSettings.flatRateMethodTitle, shipping.shippingMethod);
-			await this.clearAndType(selector.vendor.vShippingSettings.flatRateCost, shipping.shippingCost);
-			await this.selectByValue(selector.vendor.vShippingSettings.flatRateTaxStatus, shipping.taxStatus);
-			await this.clearAndType(selector.vendor.vShippingSettings.flatRateDescription, shipping.description);
-			await this.selectByValue(selector.vendor.vShippingSettings.flatRateCalculationType, shipping.calculationType);
-			break;
-
-			// free shipping
-		case 'free_shipping' :
-			await this.clearAndType(selector.vendor.vShippingSettings.freeShippingTitle, shipping.shippingMethod);
-			await this.clearAndType(selector.vendor.vShippingSettings.freeShippingMinimumOrderAmount, shipping.freeShippingMinimumOrderAmount);
-			break;
-
-			// local pickup
-		case 'local_pickup' :
-			await this.clearAndType(selector.vendor.vShippingSettings.localPickupTitle, shipping.shippingMethod);
-			await this.clearAndType(selector.vendor.vShippingSettings.localPickupCost, shipping.shippingCost);
-			await this.selectByValue(selector.vendor.vShippingSettings.localPickupTaxStatus, shipping.taxStatus);
-			await this.clearAndType(selector.vendor.vShippingSettings.flatRateDescription, shipping.description);
-			break;
-
-			// dokan table rate shipping
-		case 'dokan_table_rate_shipping' :
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingMethodTitle, shipping.shippingMethod);
-			await this.selectByValue(selector.vendor.vShippingSettings.tableRateShippingTaxStatus, shipping.taxStatus);
-			await this.selectByValue(selector.vendor.vShippingSettings.tableRateShippingTaxIncludedInShippingCosts, shipping.taxIncludedInShippingCosts);
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingHandlingFee, shipping.handlingFee);
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingMaximumShippingCost, shipping.maximumShippingCost);
-			// rates
-			// await this.selectByValue(selector.vendor.vShippingSettings.tableRateShippingCalculationType,  shipping.calculationType)
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingHandlingFeePerOrder, shipping.handlingFeePerOrder);
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingMinimumCostPerOrder, shipping.minimumCostPerOrder);
-			await this.clearAndType(selector.vendor.vShippingSettings.tableRateShippingMaximumCostPerOrder, shipping.maximumCostPerOrder);
-			await this.click(selector.vendor.vShippingSettings.tableRateShippingUpdateSettings);
-			await this.toContainText(selector.vendor.vShippingSettings.tableRateShippingUpdateSettingsSuccessMessage, shipping.tableRateSaveSuccessMessage);
-			return;
-
-			// dokan distance rate shipping
-		case 'dokan_distance_rate_shipping' :
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingMethodTitle, shipping.shippingMethod);
-			await this.selectByValue(selector.vendor.vShippingSettings.distanceRateShippingTaxStatus, shipping.taxStatus);
-			await this.selectByValue(selector.vendor.vShippingSettings.distanceRateShippingTransportationMode, shipping.transportationMode);
-			await this.selectByValue(selector.vendor.vShippingSettings.distanceRateShippingAvoid, shipping.avoid);
-			await this.selectByValue(selector.vendor.vShippingSettings.distanceRateShippingDistanceUnit, shipping.distanceUnit);
-			await this.check(selector.vendor.vShippingSettings.distanceRateShippingShowDistance);
-			await this.check(selector.vendor.vShippingSettings.distanceRateShippingShowDuration);
-			// shipping address
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingAddress1, shipping.street1);
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingAddress2, shipping.street2);
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingCity, shipping.city);
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingZipOrPostalCode, shipping.zipCode);
-			await this.clearAndType(selector.vendor.vShippingSettings.distanceRateShippingStateOrProvince, shipping.state);
-			await this.selectByValue(selector.vendor.vShippingSettings.distanceRateShippingCountry, shipping.country);
-			await this.click(selector.vendor.vShippingSettings.distanceRateShippingUpdateSettings);
-			await this.toContainText(selector.vendor.vShippingSettings.distanceRateShippingUpdateSettingsSuccessMessage, shipping.distanceRateSaveSuccessMessage);
-			return;
-
-		default :
-			break;
-		}
-		await this.clickAndWaitForResponse(data.subUrls.ajax, selector.vendor.vShippingSettings.shippingSettingsSaveSettings);
-		await this.clickAndWaitForResponse(data.subUrls.ajax, selector.vendor.vShippingSettings.saveChanges);
-		await this.toContainText(selector.vendor.vShippingSettings.updateSettingsSuccessMessage, shipping.saveSuccessMessage);
-	}
-
-
 	//todo: fixed below functions
 
 	// get total vendor earnings
@@ -919,17 +678,6 @@ export class VendorPage extends BasePage {
 
 		console.log(orderDetails);
 		return orderDetails;
-	}
-
-
-	// vendor account details render properly
-	async vendorAccountDetailsRenderProperly(){
-		await this.goIfNotThere(data.subUrls.frontend.vDashboard.editAccountVendor);
-
-		// eslint-disable-next-line @typescript-eslint/no-unused-vars
-		const { saveSuccessMessage, ...accountDetails } = selector.vendor.vAccountDetails;
-		await this.multipleElementVisible(accountDetails);
-
 	}
